--- a/web/gulpfile.ts
+++ b/web/gulpfile.ts
@@ -11,9 +11,6 @@ const WEBPACK_STATS_OPTIONS: Stats.ToStringOptions & { colors?: boolean } = {
     timings: true,
     errors: true,
     warnings: true,
-    warningsFilter: warning =>
-        // This is intended, so ignore warning
-        /node_modules\/monaco-editor\/.*\/editorSimpleWorker.js.*\n.*dependency is an expression/.test(warning),
     colors: true,
 }
 const logWebpackStats = (stats: Stats) => log(stats.toString(WEBPACK_STATS_OPTIONS))